% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_sales.R
\name{association}
\alias{association}
\title{Perform Market Basket analysis on sales transactions}
\usage{
association(code, datalake = 1)
}
\arguments{
\item{code}{Soft1 installation code}

\item{datalake}{Which Azure data lake to use (0 for production)}
}
\description{
It produces the appropriate data structures for the market basket analysis
on the sales dataset of the installation (joined with items information)
}
\details{
\itemize{
\item Must provide a valid Soft1 installation code i.e. must have data in azure data lake
\item Use 0 on datalake, for storing results on the production Azure data lake.
Use 1 (default) for the dev Azure data lake
}
}
\examples{
\dontrun{

library(tidyverse)
library(lubridate)
library(Soft1R)
library(tidyquant)
library(AzureAuth)
library(AzureStor)
library(data.table)
library(parsedate)
library(recommenderlab)
library(arules)
library(arulesViz)

association(code = "01100201021801", datalake = 1)

}

}
