% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster_data_test}
\alias{cluster_data_test}
\title{Test dataset for clustering (cluster_s1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 115 rows and 5 columns.
}
\usage{
cluster_data_test
}
\description{
This dataset was generated from a sales transactions dataset (Soft1 demo) in azure data
lake and it can be used with cluster_s1() to apply clustering
}
\examples{
cluster_data_test


}
\keyword{datasets}
