% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_s1.R
\name{cluster_s1}
\alias{cluster_s1}
\title{Apply clustering in a dataset}
\usage{
cluster_s1(final_dataset = cluster_data_all)
}
\arguments{
\item{final_dataset}{Specify the dataframe}
}
\description{
returns a dataset combining 5 separate datasets with clustering results
(for k = 2,3,4,5,6), after providing a suitable dataset
}
\details{
\itemize{
\item Must provide a dataframe with a categorical variable (first variable) and 2 or more continuous variables
\item The provided dataframe cannot include missing values
}
}
\examples{

library(tidyverse)

t <- cluster_s1(final_dataset = cluster_data_test)
t

}
