% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_dw.R
\name{connect_dw}
\alias{connect_dw}
\title{Connect to Soft1 Data warehouse}
\usage{
connect_dw(
  conn = conn,
  Driver = ifelse(base::Sys.info()["sysname"][[1]] == "Linux",
    "{ODBC Driver 17 for SQL Server}", "SQL Server"),
  Server = "analytics.softone.gr",
  Database = "ANALYTICS",
  Uid = "analytics",
  Pwd = "idlyKQcUWP\%6&$23tn"
)
}
\arguments{
\item{conn}{Specify the connection}

\item{Driver}{Specify the ODBC driver used for the connection}

\item{Server}{Specify the SQL Server of the DB}

\item{Database}{Specify the DB name}

\item{Uid}{Specify the user name}

\item{Pwd}{Specify the password}
}
\description{
Create a connection to Soft1 Data Warehouse DB, hosted in analytics.softone.gr
}
\details{
\itemize{
\item The default inputs will create a connection to the Data warehouse DB (Analytics) with full rights
}
}
\examples{
\dontrun{
library(RODBC)

Installations <- sqlQuery(connect_dw(), "Select TOP (100) * from [00_DM_Installations]")
}

}
