% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_live.R
\name{connect_live}
\alias{connect_live}
\title{Connect to Soft1 ERP database}
\usage{
connect_live(
  conn = conn,
  Driver = ifelse(base::Sys.info()["sysname"][[1]] == "Linux",
    "{ODBC Driver 17 for SQL Server}", "SQL Server"),
  Server = "11100000223803.s1data.net",
  Database = "SOFTONE",
  Uid = "analytics",
  Pwd = "idlyKQcUWP\%6&$23tn"
)
}
\arguments{
\item{conn}{Specify the connection}

\item{Driver}{Specify the ODBC driver used for the connection}

\item{Server}{Specify the SQL Server of the DB}

\item{Database}{Specify the DB name}

\item{Uid}{Specify the user name}

\item{Pwd}{Specify the password}
}
\description{
Create a connection to Soft1 ERP transactional DB, hosted in Azure
}
\details{
\itemize{
\item The default inputs will create a connection to the Soft1 DB with read rights
}
}
\examples{
\dontrun{

library(RODBC)

Installations <- sqlQuery(connect_live(), "Select TOP (100) * from [INST]")
}

}
