% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unique_ids.R
\name{create_unique_ids}
\alias{create_unique_ids}
\title{Generate unique ids}
\usage{
create_unique_ids(n, char_len = 10)
}
\arguments{
\item{n}{Specify number of ids to generate}

\item{char_len}{Specify the character length of the ID (default is 10)}
}
\description{
Generate unique ids. It can be used for generating unique keys of different lengths
}
\details{
\itemize{
\item Can be used to generate unique paths.
}
}
\examples{

create_unique_ids(5, 50)

}
