% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_sep.R
\name{decimal_sep}
\alias{decimal_sep}
\title{Replace . with ,}
\usage{
decimal_sep(input)
}
\arguments{
\item{input}{Specify the input}
}
\description{
Replace . with , It's mainly used for replacing numerical values on other regional settings
}
\details{
\itemize{
\item Mainly used for transforming decimal numbers for Greek systems.
}
}
\examples{

decimal_sep("18.8")

}
