% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_folder.R
\name{download_folder}
\alias{download_folder}
\title{Downloads specific folder & create the dataset}
\usage{
download_folder(folder)
}
\description{
At first it downloads all zip files from specific folder in SoftOne azure
datalake, perform all appropriate ETL steps (unzip, concatenate csv's, delete empty rows etc.)
& finally creates the dataset
}
\details{
\itemize{
\item It only works if there is already an authorized access to the datalake (token, installation, temp files etc. )
}
}
\examples{
\dontrun{

library(Soft1R)
library(tidyverse)
library(lubridate)

Items <- download_folder("/Items/")
Items
}

}
