% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_dataset_s1.R
\name{forecast_dataset_s1}
\alias{forecast_dataset_s1}
\title{Apply forecasting in a dataset}
\usage{
forecast_dataset_s1(
  final_dataset = vmtrstat,
  periods = "month",
  from = min(vmtrstat$date),
  to = max(vmtrstat$date),
  forecasts = 12,
  nas = "mean",
  anomaly = FALSE,
  col
)
}
\arguments{
\item{final_dataset}{Specify the dataframe}

\item{periods}{Specify the forecasting period (year, month, week, day)}

\item{from}{Specify the starting date}

\item{to}{Specify the ending date}

\item{forecasts}{Specify number of future forecasts}

\item{nas}{Specify the function to replace the missing values}

\item{anomaly}{Indicate whether or not to use anomaly detection prior to forecasting}

\item{col}{Specify the numerical variable to forecast}
}
\description{
It can generate yearly|monthly|weekly|daily forecasts, with/without anomaly detection
in one continuous variable of the dataset
}
\details{
\itemize{
\item Must provide a dataframe with at least a numerical variable (to forecast) and a date variable (must be named "date")
}
}
\examples{
\dontrun{

library(Soft1R)
library(tidyverse)
library(lubridate)

t <- forecast_dataset_s1(final_dataset = vmtrstat_test, periods = "month", from = min(vmtrstat_test$date), to = max(vmtrstat_test$date), forecasts = 12, nas = "mean", anomaly = FALSE, col = salesval)
t
}

}
