% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Push notifications with Pushover}
\usage{
notify(message)
}
\arguments{
\item{message}{Specify the message}
}
\description{
Create push notifications using the pushover application on the mobile phone
}
\details{
\itemize{
\item It will sent a push notification (printing input) to the mobile phone using pushover notification
}
}
\examples{
\dontrun{
library(pushoverr)

notify("Test notification")
}

}
