% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_various.R
\name{tukey_various}
\alias{tukey_various}
\title{Multiple Tukey tests}
\usage{
tukey_various(dataset, kpi, categorical)
}
\arguments{
\item{dataset}{Specify the dataframe}

\item{kpi}{Specify the KPI variable (Must be included in the dataframe)}

\item{categorical}{Specify a vector of categorical variables (Must be included in the dataframe)}
}
\description{
Perform multiple Tukey tests using all levels of the categorical variables
}
\details{
\itemize{
\item It returns a dataframe with multiple comparison of Tukey's range test on all levels of the categorical variables.
\item Anova returns the statistical significance of at least one level has significant difference in variance.
}
}
\examples{

library(tidyverse)

t <- tukey_various(iris, "Sepal.Length", c("Species"))
t

}
