% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmtrstat_test.R
\docType{data}
\name{vmtrstat_test}
\alias{vmtrstat_test}
\title{Test dataset for forecasting (forecast_dataset_s1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3415 rows and 19 columns.
}
\usage{
vmtrstat_test
}
\description{
This dataset was generated from a sales transactions dataset (Soft1 demo) in azure data
lake and it can be used with forecast_dataset_s1() to apply forecasting
}
\examples{
vmtrstat_test

}
\keyword{datasets}
